Move computeBlockCheckpoint to tests
